+++ clients/algoliasearch-client-php/lib/Api/AgentStudioClient.php
@@ -42,12 +42,10 @@ use GuzzleHttp\Psr7\Query;
  * AgentStudioClient Class Doc Comment.
  *
  * @category Class
- *
- * @experimental This API is in beta and subject to breaking changes. See https://www.algolia.com/doc/rest-api/agent-studio
  */
 class AgentStudioClient
 {
-    public const VERSION = '4.45.0';
+    public const VERSION = '4.46.0';
 
     /**
      * @var ApiWrapperInterface