+++ tests/output/kotlin/src/commonTest/kotlin/com/algolia/client/CompositionTest.kt
@@ -143,7 +143,7 @@ class CompositionTest {
     client.runTest(
       call = { customPost(path = "1/test") },
       intercept = {
-        val regexp = "^Algolia for Kotlin \\(3.43.0\\).*".toRegex()
+        val regexp = "^Algolia for Kotlin \\(3.45.0\\).*".toRegex()
         val header = it.headers["User-Agent"].orEmpty()
         assertTrue(
           actual = header.matches(regexp),