+++ tests/output/scala/src/test/scala/algoliasearch/client/SearchTest.scala
@@ -425,7 +425,7 @@ class SearchTest extends AnyFunSuite {
       ),
       Duration.Inf
     )
-    val regexp = """^Algolia for Scala \(2.42.0\).*""".r
+    val regexp = """^Algolia for Scala \(2.43.0\).*""".r
     val header = echo.lastResponse.get.headers("user-agent")
     assert(header.matches(regexp.regex), s"Expected $header to match the following regex: ${regexp.regex}")
   }
@@ -581,6 +581,46 @@ class SearchTest extends AnyFunSuite {
           )
         )
     }
+  }
+
+  test("deserializes null records for missing objectIDs") {
+
+    val client = SearchClient(
+      appId = "test-app-id",
+      apiKey = "test-api-key",
+      clientOptions = ClientOptions
+        .builder()
+        .withHosts(
+          List(
+            Host(
+              if (System.getenv("CI") == "true") "localhost" else "host.docker.internal",
+              Set(CallType.Read, CallType.Write),
+              "http",
+              Option(6686)
+            )
+          )
+        )
+        .build()
+    )
+
+    var res = Await.result(
+      client.getObjects(
+        getObjectsParams = GetObjectsParams(
+          requests = Seq(
+            GetObjectsRequest(
+              objectID = "foo",
+              indexName = "theIndexName"
+            ),
+            GetObjectsRequest(
+              objectID = "missing",
+              indexName = "theIndexName"
+            )
+          )
+        )
+      ),
+      Duration.Inf
+    )
+    assert(parse(write(res)) == parse("{\"results\":[{\"objectID\":\"foo\"},null]}"))
   }
 
   test("indexExists") {